Move test to other file. (#56739)
